# openapi.yaml
openapi: 3.0.3
info:
  title: ESP32 IT Tools API
  version: 1.8.2
  description: Local developer tools and device diagnostics served by ESP32-C3.
servers:
  - url: http://192.168.4.1
paths:
...
  /api/dns/lookup:
    get:
      summary: Resolve a DNS record through Debian Gateway
      parameters:
        - {name: name, in: query, required: true, schema: {type: string}}
        - {name: record_type, in: query, schema: {type: string, default: A}}
        - {name: resolver, in: query, schema: {type: string}}
      responses:
        '200': {description: DNS answers and TTL}
  /api/dns/reverse:
    get:
      summary: Reverse DNS lookup
      parameters:
        - {name: name, in: query, required: true, schema: {type: string}}
      responses:
        '200': {description: PTR answers}
  /api/dns/compare:
    get:
      summary: Compare public resolvers
      parameters:
        - {name: name, in: query, required: true, schema: {type: string}}
        - {name: record_type, in: query, schema: {type: string, default: A}}
      responses:
        '200': {description: Answers from Cloudflare, Google and Quad9}
...